(challenge2): rename challenge2 folder
