+++ README.md
@@ -1,4 +1,4 @@
-[![Build Status](https://travis-ci.org/Dubby20/Fast-Food-Fast.svg?branch=ch-integrate-travisCI-160483367)](https://travis-ci.org/Dubby20/Fast-Food-Fast)
+[![Build Status](https://travis-ci.org/Dubby20/Fast-Food-Fast.svg?branch=develop)](https://travis-ci.org/Dubby20/Fast-Food-Fast)
 [![Coverage Status](https://coveralls.io/repos/github/Dubby20/Fast-Food-Fast/badge.svg?branch=develop)](https://coveralls.io/github/Dubby20/Fast-Food-Fast?branch=develop)
 [![Maintainability](https://api.codeclimate.com/v1/badges/c5563c607a08e0628986/maintainability)](https://codeclimate.com/github/Dubby20/Fast-Food-Fast/maintainability)
 [![Test Coverage](https://api.codeclimate.com/v1/badges/c5563c607a08e0628986/test_coverage)](https://codeclimate.com/github/Dubby20/Fast-Food-Fast/test_coverage)
